Fetch: 전체 성적 평점 계산 구현

+++ Backend/src/main/java/graduationProject/graduation_judge/domain/Grade/controller/GradeController.java
@@ -2,15 +2,16 @@ package graduationProject.graduation_judge.domain.Grade.controller;
 
 import graduationProject.graduation_judge.DTO.GradeDTO;
 import graduationProject.graduation_judge.DTO.UserGradeList;
+import graduationProject.graduation_judge.DTO.UserInfoDTO;
 import graduationProject.graduation_judge.domain.Grade.service.GradeService;
+import graduationProject.graduation_judge.domain.Member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 @RestController
 @RequiredArgsConstructor
@@ -20,6 +21,9 @@ public class GradeController {
     @Autowired
     private final GradeService gradeService;
 
+    @Autowired
+    private final MemberService memberService;
+
     //성적파일 입력
     @PostMapping("/gradeFile")
     public ResponseEntity<String> inputFile(@RequestBody UserGradeList userGradeList){
@@ -28,11 +32,24 @@ public class GradeController {
             String email = userGradeList.getEmail();
             List<UserGradeList.GradeData> gradeDataList = userGradeList.getUserDataList();
 
+            // UserInfo 학기수랑 실제 파일 학기 수랑 일치하지 않으면 에러 반환
+            Set<String> uniqueTermNumbers = new HashSet<>();
+            for (UserGradeList.GradeData gradeData : gradeDataList) {
+                uniqueTermNumbers.add(gradeData.getTermNumber());
+            }
+            int uniqueTermNumberCount = uniqueTermNumbers.size(); //실제 파일 학기 수
+
+            UserInfoDTO userDTO = memberService.getMemberById(email);
+            int semUserInfo = userDTO.getSemester(); //UserInfo 학기 수
+            if (uniqueTermNumberCount != semUserInfo) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("학기 수가 일치하지 않습니다");
+            }
+
             // 기존 성적 삭제
             if(gradeService.isExistGrade(email)){
                 gradeService.deleteGradeByMember(email);
             }
-            
+
             // 새로운 성적 입력
             for (int i = 0; i < gradeDataList.size(); i++) {
                 String TNumber = gradeDataList.get(i).getTermNumber();
